Move the BestResponse out of exploitability, to be a first class citizen Policy.
We often use Best responses but we do not have any clean way of getting a `Policy` object being a BR. We often have to combine `PolicyFromCallable` with the available ways to compute a BR.

Also, the BR is computed in several places (e.g. in exploitability.py, and also in action_value_vs_best_response.py, where it uses the C++ implementation.

Having a single entry-point would make it simpler to call the C++ version whenever we discover we can, e.g. because the provided policy is a TabularPolicy.

PiperOrigin-RevId: 268238420
Change-Id: I0c11d480661263e0e9eb2a1217203d1bdc1e5f9c