Call `GetStatePolicy` with `State` instead of `std::string`.
This allows one to compute best responses for `Policy`s that do not implement `GetStatePolicy` with an info state string, e.g., non-tabular policies that choose action distributions accoring to state features. Since the default version of `GetStatePolicy` given a `State` defers to the string version, this does not require changes in any current `Policy` classes.