Change the actions in Go, from [0, 21*21] to [0, board_size * board_size].
This is a *breaking change* for Go, where the actions identifiers are being changed.
Internally, the board is a 21*21 board, and we use guard stones around the board to prevent to have to do specific checks at the border.

However, this can be unexpected for users as the action mapping is not natural. Thus, we use a mapping of the action to the given board_size and the translation is done within go.{h, cc}. Previously serialized histories will need to be rewritten.

Before this CL:
- `action` corresponds to the move 0-indexed (row, col):

  row = action // 21 - 1
  col = action % 21 - 1
- The (row, col) move is associated to the integer: (row+1) * 21 + col + 1
- Pass is 21*21.

After this CL:
- `action` corresponds to the move 0--indexed (row, col):

  row = action // board_size
  col = action % board_size.

- The (row, col) move is associated to the integer row * board_size + col.
- pass is board_size * board_size.

If you have games stored under the format used before this CL and you need to convert the list of actions to the new format, you can use:

def convert(virtual_action, board_size):
  if virtual_action = 21*21: return board_size * board_size
  row = virtual_action // 21
  col = virtual_action % 21
  return (row - 1) * board_size + col - 1

This change is error-prone. Thus, I made sure that go.txt was using the same list of actions, so we can see more easily the differences (or more precisely so that we can see there is no difference).

PiperOrigin-RevId: 294202818
Change-Id: I090a89411da5bc7fc94b02f99d8ae96bed75b231