fix current point card
When returning the current point card, the methods `ObservationString` and `ObservationTensor` look at the index of the current point card in a data structure that contains only the remaining cards, and increment it. The returned value does not always correspond to the actual value of the current point card however, especially after the first couple of moves.

One way to fix that is what I did: ObservationString and ObservationTensor now access the last element of the same vector that is also accessed by InformationStateString and InformationStateTensor, which contains the values of the point card sequence.

It is my first time contributing to a public repo. I'm sorry if I unknowingly breached some guidelines or rules.