Fixe the Go scoring algorithm.
This is an important change for Go, as the scoring system was incorrect.
During the flooding algorithm, the full connected component MUST be exhausted before we stop. We should not stop early in this process (anyway, all the cells will be checked at the end).

It's unclear the proportion of the games that are affected, but the scoring tests were passing, and we did not encounter such strange scores that often while experimenting.

Note that the error can only potentially occur when the game is not really finished and not split into full territories. When scoring a finished game, the score should always be correct.

PiperOrigin-RevId: 316632946
Change-Id: Iddf95459325010cf83eaecec8dbd4f494c267f67