Add CI testing using Github Actions.
As Travis-CI is migrating, and will remove some options for OSS projects to get free workflows, we add github actions (we can remove travis CI afterwards).

It will look like this:
https://github.com/jblespiau/open_spiel-1/actions/runs/554623616

We only have a more recent version of MacOS, with Python 3.9. The current Travis do not specify the Python 3 version, while I explicitly pin it to be Python 3.7 here.

I kept the versions the same as in .travis.yml otherwise (we have a Python 3.6, for Nox testing, we should likely upgrade it to be 3.7, but I would prefer this to be an orthogonal change.

I just have a failure for the test with Julia:
https://github.com/jblespiau/open_spiel-1/runs/1595264912?check_suite_focus=true

It should be easy to fix this using in the install script:

```
  # This is needed on Ubuntu 19.10 and above, see:
  # https://github.com/deepmind/open_spiel/issues/201
  if [[ "$JULIA_HOME" != "$HOME/packages/julias/julia-1.3.1" ]]; then
    if [[ -f /usr/lib/x86_64-linux-gnu/libstdc++.so.6 ]]; then
      echo "Julia not in $HOME. Copy libstdc++.so.6 (requires sudo)"
      sudo cp -f /usr/lib/x86_64-linux-gnu/libstdc++.so.6 $JULIA_HOME/lib/julia
```

but I simply disabled the test in this PR, and we can follow-up with that.

PiperOrigin-RevId: 356724390
Change-Id: I69141b172c73333daaefd60534376e38dc36fdea